Change floats to decimals

+++ Models/HarvestEntry.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using Newtonsoft.Json;
-using RestSharp.Serializers;
 
-namespace Paynter.Harvest.Models {
+namespace Paynter.Harvest.Models
+{
     /// <summary>
     /// Day entry given by the Harvest API. Some properties are not available in non-daily resources.
     /// </summary>
@@ -18,7 +16,7 @@ namespace Paynter.Harvest.Models {
         public int id { get; set; }
         public string notes { get; set; }
         public DateTime spent_at { get; set; }
-        public float hours { get; set; }
+        public decimal hours { get; set; }
         public int user_id { get; set; }
         public int project_id { get; set; }
         public int task_id { get; set; }
